Fix aggregate device volume slider permanently broken after install (#848)
Skip volume/mute sync for aggregate devices to prevent deprecated
AudioHardwareService APIs from corrupting their volume control state.
Always report aggregate devices as having volume/mute controls so
BGMDevice controls stay enabled and PropagateControlListChange (null
device toggle) is not triggered.

Also fix implicit int-to-float conversion warnings for newer Xcode.

+++ BGMApp/BGMApp/BGMAutoPauseMusic.mm
@@ -206,7 +206,7 @@ static Float32 const kUnpauseDelayWeightingFactor = 0.1f;
     // TODO: Fading in and out would make short pauses a lot less jarring because, if they were short enough, we wouldn't
     //       actually pause the music player. So you'd hear a dip in the music's volume rather than a gap.
     UInt64 unpauseDelayNsec =
-        static_cast<UInt64>((wentSilent - wentAudible) * kUnpauseDelayWeightingFactor);
+        static_cast<UInt64>(static_cast<Float64>(wentSilent - wentAudible) * kUnpauseDelayWeightingFactor);
     
     // Convert from absolute time to nanos.
     mach_timebase_info_data_t info;
